requirements.txt
google-adk>=0.1.0
google-genai>=0.2.0
python-dotenv>=1.0.1
mcp>=1.2.0
pydantic>=2.0.0
